+++ web/apps/photos/src/components/Collections/CollectionHeader.tsx
@@ -50,7 +50,10 @@ import { GalleryContext } from "pages/gallery";
 import React, { useCallback, useContext, useRef } from "react";
 import { Trans } from "react-i18next";
 import * as CollectionAPI from "services/collectionService";
-import * as TrashService from "services/trashService";
+import {
+    emptyTrash,
+    clearLocalTrash,
+} from "@/new/photos/services/collections";
 import { SetFilesDownloadProgressAttributesCreator } from "types/gallery";
 import {
     changeCollectionOrder,
@@ -229,13 +232,13 @@ const CollectionOptions: React.FC<CollectionOptionsProps> = ({
             continue: {
                 text: t("empty_trash"),
                 color: "critical",
-                action: emptyTrash,
+                action: doEmptyTrash,
             },
         });
 
-    const emptyTrash = wrap(async () => {
-        await TrashService.emptyTrash();
-        await TrashService.clearLocalTrash();
+    const doEmptyTrash = wrap(async () => {
+        await emptyTrash();
+        await clearLocalTrash();
         setActiveCollectionID(ALL_SECTION);
     });
